This revision contains the implementation of a Binary Search Tree, a kind of a Dictionary. The interface has methods based on the book The Algorithm Design Manual and adjusted to Java from C following my personal implementation.
The binary search tree implemented here is just an exercise. It does not support
features like self-balancing found in Red-Black Trees. The implementation
will come later (hopefully). The test cases for the traversals (In-Order, Pre-Order and Post-Order) are also provided. They all use Enums and Iterable to provide an as fluid as possible.